Refactor and migrate more functions out of api into app package
