+++ api/user_test.go
@@ -107,25 +107,6 @@ func TestCheckUserDomain(t *testing.T) {
 	}
 }
 
-func TestIsUsernameTaken(t *testing.T) {
-	th := Setup().InitBasic()
-	user := th.BasicUser
-	taken := IsUsernameTaken(user.Username)
-
-	if !taken {
-		t.Logf("the username '%v' should be taken", user.Username)
-		t.FailNow()
-	}
-
-	newUsername := "randomUsername"
-	taken = IsUsernameTaken(newUsername)
-
-	if taken {
-		t.Logf("the username '%v' should not be taken", newUsername)
-		t.FailNow()
-	}
-}
-
 func TestLogin(t *testing.T) {
 	th := Setup()
 	Client := th.CreateClient()
@@ -227,7 +208,10 @@ func TestLogin(t *testing.T) {
 	data := model.MapToJson(props)
 	hash := model.HashPassword(fmt.Sprintf("%v:%v", data, utils.Cfg.EmailSettings.InviteSalt))
 
-	ruser2, _ := Client.CreateUserFromSignup(&user2, data, hash)
+	ruser2, err := Client.CreateUserFromSignup(&user2, data, hash)
+	if err != nil {
+		t.Fatal(err)
+	}
 
 	if _, err := Client.Login(ruser2.Data.(*model.User).Email, user2.Password); err != nil {
 		t.Fatal("From verfied hash")
@@ -686,7 +670,7 @@ func TestUserCreateImage(t *testing.T) {
 	th := Setup()
 	Client := th.CreateClient()
 
-	b, err := createProfileImage("Corey Hulen", "eo1zkdr96pdj98pjmq8zy35wba")
+	b, err := app.CreateProfileImage("Corey Hulen", "eo1zkdr96pdj98pjmq8zy35wba")
 	if err != nil {
 		t.Fatal(err)
 	}
